devfusion_backend/requirements.txt
Django
djangorestframework
django-cors-headers
djangorestframework-simplejwt
Pillow